[exception][eva] feature: implement should_throw_IllegalOperationException_when_delete_given_null_employee_id

+++ spring-boot-employee/src/main/java/com/oocl/service/EmployeeService.java
@@ -58,7 +58,10 @@ public class EmployeeService {
         return foundEmployee;
     }
 
-    public void deleteByID(Integer employeeId) {
+    public void deleteByID(Integer employeeId) throws IllegalOperationException {
+        if(employeeId == null) {
+            throw new IllegalOperationException();
+        }
         employeeRepository.deleteById(employeeId);
     }
 }
